backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
python-docx==1.1.0
PyPDF2==3.0.1
openai==1.3.0
python-dotenv==1.0.0
motor==3.3.2
pydantic==2.5.0
scikit-learn==1.3.2
numpy==1.25.2
google-generativeai